Don't check if module is built-in
This check is harmful since the set of doc pages is not in one-to-one
correspondence with the list of builtins.

Hence this check would require keeping list of doc pages in package
in sync with actual Node docs or sending another HTTP request just
to get this list, which is absurd.